style-changes-all

+++ src/ImageGallery/ImageGalleryRowCard.js
@@ -4,18 +4,41 @@ import { withRouter, useHistory } from "react-router-dom";
 import Tooltip from "@material-ui/core/Tooltip";
 import EditIcon from "@material-ui/icons/Edit";
 import DeleteIcon from "@material-ui/icons/Delete";
+import Swal from 'sweetalert2'
 
 function ImageGalleryRowCard(props) {
   const history = useHistory();
+
+  function confirmDelete(){
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        props.deleteImageGallery(props.id);
+        Swal.fire(
+          'Deleted!',
+          'The image has been deleted.',
+          'success'
+        )
+      }
+    })
+  }
 
   const showTool = props.isAuth && (
     <div className="flex flex-row justify-evenly w-full ml-3 ">
       <Tooltip title="Delete Image Gallery">
         <DeleteIcon
-          onClick={() => {
-            props.deleteImageGallery(props.id);
-          }}
-          className="material-icons  cursor-pointer  transform hover:scale-110 motion-reduce:transform-none group-hover:text-black"
+          // onClick={() => {
+          //   props.deleteImageGallery(props.id);
+          // }}
+          onClick={confirmDelete}
+          className="p-1 rounded-full bg-pink-900  mb-1 group-hover:text-white material-icons  cursor-pointer transform hover:scale-110 motion-reduce:transform-none"
         />
       </Tooltip>
 
@@ -24,7 +47,7 @@ function ImageGalleryRowCard(props) {
           onClick={() => {
             props.editView(props.id);
           }}
-          className="material-icons  cursor-pointer  transform hover:scale-110 motion-reduce:transform-none group-hover:text-black"
+          className="p-1 rounded-full formBG  mb-1 group-hover:text-white material-icons  cursor-pointer transform hover:scale-110 motion-reduce:transform-none"
         />
       </Tooltip>
     </div>
@@ -41,8 +64,8 @@ function ImageGalleryRowCard(props) {
   );
 
   return (
-    <Card className="ml-3 mr-3  shadow w-80 md:w-96">
-      <div className=" p-2 flex flex-row justify-between text-center bg-gray-700 text-gray-300 hover:bg-gray-50 hover:border-transparent hover:shadow-xl group cursor-pointer">
+    <Card className="ml-3 mr-3  shadow w-80 md:w-96 border-0">
+      <div className=" p-2 flex flex-row justify-between text-center formBG text-gray-300 hover:bg-gray-50 hover:border-transparent hover:shadow-xl group cursor-pointer">
         <div>{showTool}</div>
 
         {/* called from ImageGalleryIndex  - Dispay Movie-Drama Title*/}
@@ -54,7 +77,7 @@ function ImageGalleryRowCard(props) {
                   -1 && (
                   <Tooltip title={`Go to ${md.title} ${md.type} details page`}>
                     <Card.Text
-                      className="group-hover:text-gray-800 cursor-pointer"
+                      className="group-hover:text-gray-200 cursor-pointer"
                       onClick={() => {
                         history.push({
                           pathname: `/movieDramaDetails/${md.title}`,
@@ -70,13 +93,6 @@ function ImageGalleryRowCard(props) {
             ))}
           </div>
         )}
-
-        {/* called from MDDetails  - Dispay Movie-Drama Title*/}
-        {/* {props.movieDrama && (
-          <div className="group-hover:text-gray-800 w-full text-center">
-            {props.movieDrama.title}
-          </div>
-        )} */}
       </div>
 
       {cardImage}
